Setup sbt build and convert code to scala
Converted major idioms to scala from java. I had to leave out the code
to read from a file and return values based on that.

I also have not yet implemented logging yet. I will try to get that done
as well.

Finally I configured the scala build tool to be able to build and manage
the akka dependency.